+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router": "^6.20.0",
     "react-router-dom": "^6.16.0"
   },
   "devDependencies": {
